--- src/helper/idParsers.ts
@@ -1,3 +1,0 @@
-export const mentionChannelWithId = (channelId: string): string => {
-    return `<#${channelId}>`;
-}
